update on user progress and management
