+++ app/api/auth/[...nextauth]/route.ts
@@ -1,74 +1,6 @@
-import NextAuth, { NextAuthOptions } from "next-auth";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import CredentialsProvider from "next-auth/providers/credentials";
-import GoogleProvider from "next-auth/providers/google";
-import { prisma } from "@/lib/prisma";
-import { compare } from "bcrypt";
-export const authOptions: NextAuthOptions = {
-	adapter: PrismaAdapter(prisma),
-	session: {
-		strategy: "jwt" as const,
-	},
-	providers: [
-		CredentialsProvider({
-			name: "Credentials",
-			credentials: {
-				email: { label: "Email", type: "email" },
-				password: { label: "Password", type: "password" },
-			},
-			async authorize(credentials) {
-				if (!credentials?.email || !credentials?.password) return null;
-				const user = await prisma.user.findUnique({
-					where: { email: credentials.email },
-				});
-				if (!user) return null;
-
-				const isValid = await compare(credentials.password, user.password || "");
-				if (!isValid) return null;
-
-				return { id: user.id, name: user.name, email: user.email };
-			},
-		}),
-
-		GoogleProvider({
-			clientId: process.env.GOOGLE_CLIENT_ID!,
-			clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-			allowDangerousEmailAccountLinking: true,
-		}),
-	],
-	callbacks: {
-		async jwt({ token, user }) {
-			if (user) {
-				token.id = user.id;
-			}
-			return token;
-		},
-		async signIn({ user, account }) {
-			if (account?.provider === "google") {
-				if (!user.email) return false; // 👈 check added here
-
-				const existingUser = await prisma.user.findUnique({
-					where: { email: user.email },
-				});
-
-				if (!existingUser) {
-					await prisma.user.create({
-						data: {
-							email: user.email,
-							name: user.name,
-							image: user.image,
-						},
-					});
-				}
-			}
-			return true;
-		},
-	},
-	pages: {
-		signIn: "/signin", // Optional: custom sign in page
-		error: "/signin",
-	},
-};
+import NextAuth from "next-auth";
+import { authOptions } from "@/lib/auth";
 
 const handler = NextAuth(authOptions);
+
 export { handler as GET, handler as POST };
